requirements.txt
cfgv==3.5.0
distlib==0.4.0
filelock==3.25.0
identify==2.6.17
nodeenv==1.10.0
noise==1.2.2
pathfinding==1.0.20
platformdirs==4.9.4
pre_commit==4.5.1
pygame-ce==2.5.7
python-discovery==1.1.0
PyYAML==6.0.3
virtualenv==21.1.0
